Move make_cluster_fixture to test_util
